pyproject.toml
[project]
name = "forecast-aisafety"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "jupyter>=1.1.1",
    "numpy>=2.3.2",
    "openai>=1.99.9",
    "pandas>=2.3.1",
    "plotly>=6.3.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.48.1",
]
